Add files via upload
ICs for the detection of task stages. - A) Independent Component corresponding to the object presentation, note that in some trials the characteristic peak of the IC is visible, but its amplitude does not exceed the set detection threshold. B) Precision and Recall for first monkey "S".  Note the greater variability in the detection of the planning stages "Object Presentation" and "Grip Cue", from "Go Cue" to the end of the task, values above 90% are reached. C) Stage detection results for monkey "R". Error bars reflect the standard deviation (SD) across the five recording sessions for each monkey.