Add files via upload
Relative contribution of each cortical area to LFP ICs assessed using the method described in Fig.5A. Each colored asterisk represents the average value of the ICA weights in each area, in each stage of the movement and in each recording session, relative to the expected chance distribution in orange. For example, the asterisks in cyan represent the Begin_L component, it can be seen that for this ICA component the MI contribution is more than the expected value by chance, in PMv it is lower and in PMd it is within the range expected by chance. This pattern is similar in the five recordings. The blue dotted lines represent the 99% confidence intervals