Add files via upload
Figure 7.- LFP ICs contain information related to target object and grasping strategy. An LDA classifier with 70% training and 30% test data was used to predict object type and grasping strategy. Features were extracted from the low frequency LFP band using a 100ms sliding window with 50% overlap. The feature vector was created using the average amplitude of each IC. The results for each monkey are shown separately (mean of 5 sessions for each animal). The standard deviation is presented with a shaded line obtained with 200 runs. The shaded horizontal bar represents the probability of chance +/- one standard deviation. Panels G and H show the histogram of the significance of the predictors, the probability distribution suggests.